+++ thymus/tui/__init__.py
@@ -2,6 +2,4 @@ __all__ = (
     'OpenDialog',
 )
 
-__version__ = '0.1.1'
-
 from .open_dialog import OpenDialog