+++ thymus/tuier.py
@@ -7,12 +7,14 @@ from textual.widgets import (
     TextLog,
     ListView,
 )
-from textual.containers import Horizontal
+from textual.containers import Horizontal, Vertical
 from textual.reactive import var
+from rich.text import Text
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Optional
 
-from . import __version__ as appver
+from . import __version__ as app_ver
+from .app_settings import AppSettings
 from .tui import OpenDialog
 
 
@@ -21,7 +23,7 @@ if TYPE_CHECKING:
     from textual.screen import Screen
 
 
-class MExtendedListView(ListView):
+class MExtendedListView(ListView, can_focus=True):
     def action_select_cursor(self) -> None:
         if selected := self.highlighted_child:
             if value := selected.children[0].name:
@@ -38,19 +40,33 @@ class TThymus(App):
         ('ctrl+d', 'dark_mode', 'Toggle dark mode'),
         ('ctrl+s', 'main_screen', 'Switch to main'),
     ]
-    default_screen: var['Screen | None'] = var(None)
+    default_screen: var['Optional[Screen]'] = var(None)
+    settings: var[AppSettings] = var(AppSettings())
+
+    def __log_error(self, error: str) -> None:
+        if error and (log := self.query_one('#main-app-log', TextLog)):
+            log.write(Text(error, style='red'))
 
     def compose(self) -> 'ComposeResult':
         yield Footer()
-        yield Static(f'Thymus ver. {appver}', id='main-welcome-out')
+        yield Static(Text(f'Thymus ver. {app_ver}.', style='green italic'), id='main-welcome-out')
         yield Horizontal(
-            TextLog(id='main-app-log'),
-            MExtendedListView(id='main-screens-section'),
+            Vertical(
+                Static('Application log:'),
+                TextLog(id='main-app-log'),
+            ),
+            Vertical(
+                Static('Open contexts list (select one and press Enter):'),
+                MExtendedListView(id='main-screens-section'),
+            ),
             id='main-middle-container'
         )
 
     def on_compose(self) -> None:
         self.default_screen = self.screen
+
+    def on_ready(self) -> None:
+        self.settings.playback(self.__log_error)
 
     def action_main_screen(self) -> None:
         if self.default_screen: