+++ release_notes.md
@@ -3,3 +3,14 @@
 * Textual support updated to the latest version (v0.24.1).
 * The open dialog was redesigned.
 * An extended textlog widged's code was simplified with the scroll_end option.
+
+# Version 0.1.2-alpha
+
+* The main menu was redesigned.
+* Thymus supports settings and the settings file now.
+* User can manipulate the settings via the `global` commands.
+* Syntax highlighting.
+* The `up show` modificator for the TUI part.
+* Fixed bug with double rows of the TreeView in the open dialog.
+* Fixed bug with the Input in the open dialog, it supports Enter hit now.
+* Fixed bug with ths save modificator, it does not crash the app now.